+++ cfgov/v1/models/__init__.py
@@ -6,9 +6,6 @@ from v1.models.base import (
     CFGOVPage,
     CFGOVPageCategory,
     CFGOVTaggedPages,
-    FailedLoginAttempt,
-    PasswordHistoryItem,
-    TemporaryLockout,
 )
 from v1.models.blog_page import BlogPage, LegacyBlogPage
 from v1.models.browse_filterable_page import (
